Refactor Employee editing to use new view model
Updated the `EmployeeController` to utilize `EmployeeDepartmentListViewModel` for the `Edit` and `SaveEdit` actions. This change allows for the inclusion of a department list in the view model, enhancing the edit form with a dropdown for department selection. The `Edit.cshtml` view has been modified accordingly, and a new view model class has been created to support these changes.